pyproject.toml
[project]
name = "chatterbox-tts-api"
version = "1.3.0"
description = "REST API for Chatterbox TTS with OpenAI compatibility"
authors = [{ name = "Travis Van Nimwegen", email = "tts@travis2.com" }]
readme = "README.md"
license = { text = "AGPLv3" }
requires-python = ">=3.11"
dependencies = [
  "chatterbox-tts @ git+https://github.com/resemble-ai/chatterbox.git",
  "resemble-perth",
  "torch>=2.0.0,<2.7.0",
  "torchaudio>=2.0.0,<2.7.0",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "pydantic>=2.0.0",
  "python-multipart>=0.0.6",
  "python-dotenv>=1.0.0",
  "psutil>=5.9.0",
  "requests>=2.28.0",
]

[project.urls]
Homepage = "https://github.com/travisvn/chatterbox-tts-api"
Repository = "https://github.com/travisvn/chatterbox-tts-api"

[project.optional-dependencies]
dev = [
  "requests>=2.28.0", # for testing
]

[tool.uv]
dev-dependencies = ["requests>=2.28.0"]
